# Is<particle_type> cross checks and caching