--- a/meta/TreeMasses.m
+++ b/meta/TreeMasses.m
@@ -355,45 +355,50 @@ IsSMParticleElementwise[sym_, OptionsPattern[]] :=
               Table[sym[i], {i, GetDimension[sym]}]
     ];
 
-IsScalar[Susyno`LieGroups`conj[sym_]] := IsScalar[sym];
 IsScalar[SARAH`bar[sym_]] := IsScalar[sym];
 Module[{cache},
+   IsScalar[Susyno`LieGroups`conj[sym_]] := cache[sym];
    IsScalar[sym_] := cache[sym];
+
    cache[sym_] := cache[sym] = IsOfType[sym, S];
 ];
 IsScalar[sym_List] := And @@ (IsScalar /@ sym);
 IsScalar // Utils`MakeUnknownInputDefinition;
 
 IsFermion[Susyno`LieGroups`conj[sym_]] := IsFermion[sym];
-IsFermion[SARAH`bar[sym_]] := IsFermion[sym];
 Module[{cache},
+   IsFermion[SARAH`bar[sym_]] := cache[sym];
    IsFermion[sym_] := cache[sym];
+
    cache[sym_] := cache[sym] = IsOfType[sym, F];
 ];
 IsFermion[sym_List] := And @@ (IsFermion /@ sym);
 IsFermion // Utils`MakeUnknownInputDefinition;
 
-IsVector[Susyno`LieGroups`conj[sym_]] := IsVector[sym];
 IsVector[SARAH`bar[sym_]] := IsVector[sym];
 Module[{cache},
+   IsVector[Susyno`LieGroups`conj[sym_]] := cache[sym];
    IsVector[sym_] := cache[sym];
+
    cache[sym_] := cache[sym] = IsOfType[sym, V];
 ];
 IsVector[sym_List] := And @@ (IsVector /@ sym);
 IsVector // Utils`MakeUnknownInputDefinition;
 
 IsGhost[Susyno`LieGroups`conj[sym_]] := IsGhost[sym];
-IsGhost[SARAH`bar[sym_]] := IsGhost[sym];
 Module[{cache},
+   IsGhost[SARAH`bar[sym_]] := cache[sym];
    IsGhost[sym_] := cache[sym];
+
    cache[sym_] := cache[sym] = IsOfType[sym, G];
 ];
 IsGhost[sym_List] := And @@ (IsGhost /@ sym);
 
-IsGoldstone[Susyno`LieGroups`conj[sym_]] := IsGoldstone[sym];
 IsGoldstone[SARAH`bar[sym_]] := IsGoldstone[sym];
 Module[{cache},
    IsGoldstone[sym_] := cache[sym];
+   IsGoldstone[Susyno`LieGroups`conj[sym_]] := cache[sym];
+
    cache[sym_] := cache[sym] = MemberQ[
        Join[GetGoldstoneBosons[],
             GetGoldstoneBosons[] /. a_[{idx__}] :> a[idx]],