* Improve the speed for multiple calls thanks to caching.
* Print a warning if unexpected input is parsed.