@wkotlarski 2021-11-17
I have 2 questions:

1. Do those functions actually use so much time that this complication is justified.
2. Why not just make functions remember computed values (see [here](https://reference.wolfram.com/language/workflow/WriteAFunctionThatRemembersComputedValues.html))?