@uukhas 2021-11-17
Let me start from this question:
>     2. And if yes, why not just make functions remember computed values (see [here](https://reference.wolfram.com/language/workflow/WriteAFunctionThatRemembersComputedValues.html))?
I am using exactly this mechanism, but for `cache$***` function:
```mathematica
Module[{cache},
   IsFermion[SARAH`bar[sym_]] := cache[sym];
   IsFermion[sym_] := cache[sym];
   cache[sym_] := cache[sym] = IsOfType[sym, F];
];
IsFermion[sym_List] := And @@ (IsFermion /@ sym);
IsFermion // Utils`MakeUnknownInputDefinition;
IsFermion // Protected;
```
Why? Because of the line `IsFermion // Protected;`, which forbids me to do it via "normal" `IsFermion[bar[sym_]] := IsFermion[bar@sym] = <...>`. 

Why `Protected`? Because I would like to prevent some unexpected re-definitions somewhere in the other parts of the code.

Also, in order to prevent ill-defined usage I use ``IsFermion // Utils`MakeUnknownInputDefinition;`` as well.