@uukhas 2021-11-17
>     1. Do those functions actually use so much time that this complication is justified?
**So much time**? - for `SM` and `MSSM` not really (typically ~20%):
```mathematica
IsScalar@bar@hh; (* Call with message: first call 2e-2s, the next ones 1e-5s *)
IsScalar@hh; (* Call without message: first call 5e-6s, the next ones ~4e-6s *)
```