@wkotlarski 2021-11-17
Can you show any measurable difference in runtime for the user?