@uukhas 2021-11-17
@wkotlarski, for `SM`:
|| old  | new | boost|
| --- | --- | --- | --- |
| real | 1m 17.515s  | 1m 17.**166**s  | **0.5**% |
| user | 2m 5**2.974**s  | 2m 53.538s  | -0.2% |
| sys | 8.438s  | 8.**312**s  | **1.5**% |

The code does not make the calculation slower or faster at this point. What it does is:
a) Some *almost* free physical checks
b) Speed up in future, if `Is<...>` functions are used more often