@wkotlarski 2021-11-17
My point is that we're introducing a lot of conceptual complications for no gain. Before it was obvious what those functions were doing. Now it is not. I'm sorry, but I fail to see benefits in this change.