@uukhas 2021-11-17
Benefit is that if there is a `bar@VP`, for example, somewhere in the code, then `FlexibleSUSY` can at least warn the user, that something might go wrong. Also, it is still obvious what the function is doing, because of its name.