@uukhas 2021-11-17
You use, for example, `std::cout` because you have a simple name and a description. You do not care, how the actual realization of the function is done.