@uukhas 2021-11-17
The point is that I did not change the interface of this function, but I added some nice physics-program-specific checks with no loss of the speed and with some benefits in scaling. Why should this improvement be discarded?