@wkotlarski 2021-11-17
You're talking from the point of view of a user. But someone will come after you and will have to maintain it. I'm sorry, but my opinion is that the caching is an unneeded complication. But it is @Expander that decides what too do. I just voiced my opinion :heart: