@uukhas 2021-11-17
But this is an improvement in maintenance!

a) Someone can create a weird ``SARAH` `` file. We do not want do know, how this external package works. But what we know, is that we do not want `bar@VP` or  `bar@hh`, or `conj@Fe` in the code, because it is not intuitive. So, we check indirectly some inputs.

b) Someone in the future would be creating some ``FlexibleSUSY` `` code. And this person would like to create something like `bar@VP` by wish or by accident. And my improvement has a chance to warn about this, is `Is<...>` is called.

c) Now, if someone calls `IsVector[SOMEHEAD[Vp, Fe, gG]]`, then the code is just not evaluated, nothing more. It is not working if `SOMEHEAD` is not `List`. And this might lead to a many-hours debugging. Because of ``Utils`MakeUnknowInputDefinitions`` our `FlexibleSUSY` will immediately stop and warn the user about this issue.

d) `IsScalar` - is a very nice name! Short and self descriptive. This is why someone will definitely create the same function in the future a well. And, because of default behavior of `Mathematica` for contexts, this new definition will be added or even override some existing ones in `TreeMasses`. Now, it is not possible, because symbols are `Protected`.

4 benefits for free.