@wkotlarski 2021-11-17
BTW, there's something that came to my mind. If it's about solving the problem with `conj` not a scalar etc, wouldn't it be better to move it to some other place? Because in principle we'd now had to do the same for every other function that might take a field wrapped in a head. Wouldn't it make more sense to make a rule that `conj` and `bar` check their arguments automatically?