@uukhas 2021-11-17
@wkotlarski, yes. I absolutely agree. Where should we do this instead? For `VVVV` vertices it was possible to make a wrapper in the PR relevant for #271. Here, it might be a bit different, because It seems that `bar` and `conj` are already really strongly specialized, so that it is hard for me to imagine, how to do this. There might be better to create a separate, orthogonal. Checker, maybe in the `Wrappers`.