@uukhas 2021-11-17
@wkotlarski, you are probably right after all.

This improvement shall be done only when it is required. So far there were no reports about the problem, which I tried to solve.